refactor: move v1 and v1beta1 only things to separate files
