+++ dataclients/kubernetes/definitions/doc.go
@@ -1,0 +1,3 @@
+/* Package definitions provides type definitions, parsing, marshaling and
+validation for Kubernetes resources used by Skipper. */
+package definitions
