+++ dataclients/kubernetes/definitions/ingressv1.go
@@ -2,12 +2,21 @@ package definitions
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
 	"strconv"
+
+	"github.com/pkg/errors"
+	"gopkg.in/yaml.v2"
 )
 
-var errInvalidPortType = errors.New("invalid port type")
+type IngressV1List struct {
+	Items []*IngressV1Item `json:"items"`
+}
+
+type IngressV1Item struct {
+	Metadata *Metadata      `json:"metadata"`
+	Spec     *IngressV1Spec `json:"spec"`
+}
 
 // IngressSpecV1 https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.22/#ingressspec-v1-networking-k8s-io
 type IngressV1Spec struct {
@@ -62,94 +71,43 @@ type PathRuleV1 struct {
 	Backend  *BackendV1 `json:"backend"`
 }
 
-// IngressSpec is the v1beta1
-type IngressSpec struct {
-	DefaultBackend *Backend `json:"backend"`
-	Rules          []*Rule  `json:"rules"`
+// ParseIngressV1JSON parse JSON into an IngressV1List
+func ParseIngressV1JSON(d []byte) (IngressV1List, error) {
+	var il IngressV1List
+	err := json.Unmarshal(d, &il)
+	return il, err
 }
 
-type Backend struct {
-	ServiceName string      `json:"serviceName"`
-	ServicePort BackendPort `json:"servicePort"`
-	// Traffic field used for custom traffic weights, but not part of the ingress spec.
-	Traffic float64
-	// number of True predicates to add to support multi color traffic switching
-	NoopCount int
+// ParseIngressV1YAML parse YAML into an IngressV1List
+func ParseIngressV1YAML(d []byte) (IngressV1List, error) {
+	var il IngressV1List
+	err := yaml.Unmarshal(d, &il)
+	return il, err
 }
 
-type Rule struct {
-	Host string    `json:"host"`
-	Http *HTTPRule `json:"http"`
-}
-
-type BackendPort struct {
-	Value interface{}
-}
-
-type HTTPRule struct {
-	Paths []*PathRule `json:"paths"`
-}
-
-type PathRule struct {
-	Path    string   `json:"path"`
-	Backend *Backend `json:"backend"`
-}
-
-type ResourceID struct {
-	Namespace string
-	Name      string
-}
-
-func (b Backend) String() string {
-	return fmt.Sprintf("svc(%s, %s) %0.2f", b.ServiceName, b.ServicePort, b.Traffic)
-}
-
-func (p BackendPort) Name() (string, bool) {
-	s, ok := p.Value.(string)
-	return s, ok
-}
-
-func (p BackendPort) Number() (int, bool) {
-	i, ok := p.Value.(int)
-	return i, ok
-}
-
-func (p *BackendPort) UnmarshalJSON(value []byte) error {
-	if value[0] == '"' {
-		var s string
-		if err := json.Unmarshal(value, &s); err != nil {
-			return err
-		}
-
-		p.Value = s
-		return nil
-	}
-
-	var i int
-	if err := json.Unmarshal(value, &i); err != nil {
-		return err
-	}
-
-	p.Value = i
+// TODO: implement once IngressItem has a validate method
+// ValidateIngressV1 is a no-op
+func ValidateIngressV1(_ *IngressV1Item) error {
 	return nil
 }
 
-func (p BackendPort) MarshalJSON() ([]byte, error) {
-	switch p.Value.(type) {
-	case string, int:
-		return json.Marshal(p.Value)
-	default:
-		return nil, errInvalidPortType
+// ValidateIngresses is a no-op
+func ValidateIngressesV1(ingressList IngressV1List) error {
+	var err error
+	// discover all errors to avoid the user having to repeatedly validate
+	for _, i := range ingressList.Items {
+		nerr := ValidateIngressV1(i)
+		if nerr != nil {
+			name := i.Metadata.Name
+			namespace := i.Metadata.Namespace
+			nerr = fmt.Errorf("%s/%s: %w", name, namespace, nerr)
+			err = errors.Wrap(err, nerr.Error())
+		}
 	}
-}
 
-func (p BackendPort) String() string {
-	switch v := p.Value.(type) {
-	case string:
-		return v
-	case int:
-		return strconv.Itoa(v)
-	default:
-		return ""
+	if err != nil {
+		return err
 	}
+
+	return nil
 }
